Updating template layout and documenting service
